requirements.txt
# Web Framework and Extensions
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==24.0.0

# Task Queue
celery==5.6.2
redis==7.1.0

# Data Validation and Environment
pydantic==2.12.5
email-validator==2.3.0
python-dotenv==1.2.1

# Database Driver
psycopg2-binary==2.9.11

# Data Science and Machine Learning
pandas==3.0.0
transformers==4.57.6
torch==2.10.0

# Networking and Utilities
requests==2.32.5

# Testing
pytest==9.0.2
pytest-cov==7.0.0